--- aries-site/src/examples/templates/PageHeaderExample.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { Header, Heading, Button } from 'grommet';
-
-export const PageHeaderExample = ({ title }) => (
-  <Header>
-    <Heading margin={{ vertical: 'medium' }} size="small">
-      {title}
-    </Heading>
-    <Button label="Manage" primary />
-  </Header>
-);
-
-PageHeaderExample.propTypes = {
-  title: PropTypes.string,
-};
