--- aries-site/src/examples/templates/ScreenContainer.js
@@ -1,36 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { Box, ResponsiveContext } from 'grommet';
-import { SidebarExample } from '.';
-
-export const ScreenContainer = ({ mobile, ...rest }) => (
-  <ResponsiveContext.Consumer>
-    {size => (
-      <Box
-        background="background-back"
-        width={size === 'small' ? 'medium' : '100%'}
-        height={size === 'small' ? { max: 'large' } : undefined}
-        style={{ position: 'relative' }}
-      >
-        <Box direction="row" fill>
-          {size !== 'small' && <SidebarExample />}
-          <Box
-            overflow="auto"
-            pad={{ horizontal: 'medium', bottom: 'medium' }}
-            flex
-            {...rest}
-          />
-        </Box>
-        {size === 'small' && (
-          <SidebarExample
-            style={{ position: 'absolute', bottom: 0, left: 0 }}
-          />
-        )}
-      </Box>
-    )}
-  </ResponsiveContext.Consumer>
-);
-
-ScreenContainer.propTypes = {
-  mobile: PropTypes.bool,
-};
