+++ aries-site/src/examples/templates/dashboards/index.js
@@ -1,2 +1,1 @@
 export * from './DashboardExample';
-export * from './DashboardTiles';
