+++ aries-site/src/examples/templates/forms/SimpleSignUpExample.js
@@ -13,8 +13,80 @@ import {
   TextInput,
 } from 'grommet';
 
-import { FormContainer } from '.';
-import { emailMask, emailValidation, passwordRulesStrong } from './formHelpers';
+const passwordRulesStrong = [
+  {
+    regexp: new RegExp('(?=.*?[A-Z])'),
+    message: 'At least one uppercase English letter',
+    status: 'error',
+  },
+  {
+    regexp: new RegExp('(?=.*?[a-z])'),
+    message: 'At least one lowercase English letter',
+    status: 'error',
+  },
+  {
+    regexp: new RegExp('(?=.*?[0-9])'),
+    message: 'At least one number',
+    status: 'error',
+  },
+  {
+    regexp: new RegExp('(?=.*?[#?!@$ %^&*-])'),
+    message: 'At least one special character or space',
+    status: 'error',
+  },
+  {
+    regexp: new RegExp('.{8,}'),
+    message: 'At least eight characters',
+    status: 'error',
+  },
+];
+
+const emailValidation = [
+  {
+    regexp: new RegExp('[^@ \\t\\r\\n]+@'),
+    message: 'Missing an @?',
+    status: 'info',
+  },
+  {
+    regexp: new RegExp('[^@ \\t\\r\\n]+@[^@ \\t\\r\\n]+\\.[^@ \\t\\r\\n]+'),
+    message: 'Missing an .?',
+    status: 'info',
+  },
+  {
+    regexp: new RegExp('[^@ \\t\\r\\n]+@[^@ \\t\\r\\n]+\\.[^@ \\t\\r\\n]+'),
+    message: "Email address doesn't look quite right",
+    status: 'error',
+  },
+];
+
+const emailMask = [
+  {
+    regexp: /^[\w\-_.]+$/,
+    placeholder: 'jane.smith',
+  },
+  { fixed: '@' },
+  {
+    regexp: /^[\w]+$/,
+    placeholder: 'hpe',
+  },
+  { fixed: '.' },
+  {
+    regexp: /^[\w]+$/,
+    placeholder: 'com',
+  },
+];
+
+const FormContainer = ({ ...rest }) => {
+  return (
+    <Box background="background-front" border round="small" overflow="hidden">
+      <Box
+        flex
+        pad={{ horizontal: 'medium', vertical: 'medium' }}
+        {...rest}
+       />
+    </Box>
+  );
+};
 
 export const SimpleSignUpExample = () => {
   const [formValues, setFormValues] = React.useState({
