+++ aries-site/src/examples/templates/forms/index.js
@@ -2,7 +2,6 @@ export * from './ChangePasswordExample';
 export * from './CustomizeExample';
 export * from './FilterExample';
 export * from './FormContainer';
-export * from './formHelpers';
 export * from './PayExample';
 export * from './SettingsExample';
 export * from './ShippingExample';
