+++ aries-site/src/examples/templates/index.js
@@ -1,7 +1,3 @@
-export * from './AppHeaderExample';
 export * from './dashboards';
 export * from './forms';
 export * from './list-views';
-export * from './PageHeaderExample';
-export * from './ScreenContainer';
-export * from './SidebarExample';
