+++ aries-site/src/examples/templates/list-views/ListScreenExample.js
@@ -197,33 +197,30 @@ const AppHeaderExample = () => (
   </Header>
 );
 
-const ScreenContainer = ({ mobile, ...rest }) => (
-  <ResponsiveContext.Consumer>
-    {size => (
-      <Box
-        background="background-back"
-        width={size === 'small' ? 'medium' : '100%'}
-        height={size === 'small' ? { max: 'large' } : undefined}
-        style={{ position: 'relative' }}
-      >
-        <Box direction="row" fill>
-          {size !== 'small' && <SidebarExample />}
-          <Box
-            overflow="auto"
-            pad={{ horizontal: 'medium', bottom: 'medium' }}
-            flex
-            {...rest}
-          />
-        </Box>
-        {size === 'small' && (
-          <SidebarExample
-            style={{ position: 'absolute', bottom: 0, left: 0 }}
-          />
-        )}
+const ScreenContainer = ({ mobile, ...rest }) => {
+  const size = useContext(ResponsiveContext);
+  return (
+    <Box
+      background="background-back"
+      width={size === 'small' ? 'medium' : '100%'}
+      height={size === 'small' ? { max: 'large' } : undefined}
+      style={{ position: 'relative' }}
+    >
+      <Box direction="row" fill>
+        {size !== 'small' && <SidebarExample />}
+        <Box
+          overflow="auto"
+          pad={{ horizontal: 'medium', bottom: 'medium' }}
+          flex
+          {...rest}
+        />
       </Box>
-    )}
-  </ResponsiveContext.Consumer>
-);
+      {size === 'small' && (
+        <SidebarExample style={{ position: 'absolute', bottom: 0, left: 0 }} />
+      )}
+    </Box>
+  );
+};
 
 export const ListScreenExample = ({ ...rest }) => {
   return (
