+++ aries-site/src/examples/templates/list-views/index.js
@@ -1,2 +1,1 @@
 export * from './ListScreenExample';
-export * from './StyledList';
